Fixed assign_fakeout_type_2D for recentering trials (again)
In the most recent sessions for SC_Ephys_4 - 6, we now have the option of making recentering trials either the same as the center position in both AP/ML, the same as the center position in ML with different AP, or different ML (due to user error).  The new version takes this into account.